requirements.txt
# OAuth Authentication
django-allauth==0.60.1
requests-oauthlib==1.3.1

# Existing packages (if any)
Django>=5.2.4
channels>=4.0.0
djangorestframework>=3.14.0
